+++ lib/src/data/drivers/firebase/firebase_analytics_driver.dart
@@ -24,12 +24,18 @@ class FirebaseAnalyticsDriver extends IFirebaseAnalyticsDriver {
     }
   }
 
-  Map<String, Object> _convertToMapStringObject(Map params) {
+  Map<String, Object>? _convertToMapStringObject(Map? params) {
+    if (params == null) return null;
     return params.map((key, value) {
       if (value is String || value is num) {
         return MapEntry(key, value);
       } else {
-        return MapEntry(key, value.toString().substring(0, 100));
+        return MapEntry(
+          key,
+          value.toString().length > 100
+              ? value.toString().substring(0, 100)
+              : value.toString(),
+        );
       }
     });
   }
@@ -43,7 +49,7 @@ class FirebaseAnalyticsDriver extends IFirebaseAnalyticsDriver {
       await instance.logEvent(
         name: event.name,
         callOptions: AnalyticsCallOptions(global: true),
-        parameters: _convertToMapStringObject(event.parameters ?? {}),
+        parameters: _convertToMapStringObject(event.parameters),
       );
       debugPrint(
         '''>>> FirebaseAnalyticsDriver.createEvent [${event.name}] ${event.parameters} <<<''',
@@ -65,9 +71,9 @@ class FirebaseAnalyticsDriver extends IFirebaseAnalyticsDriver {
   }) async {
     try {
       await Future.wait([
-        instance.logLogin(loginMethod: loginMethod, parameters: params),
         instance.setUserId(id: value),
         setUserProperty(name: name, value: value),
+        instance.logLogin(loginMethod: loginMethod, parameters: params),
       ]);
       for (MapEntry<String, Object> item in params?.entries ?? []) {
         setUserProperty(name: item.key, value: item.value.toString());
@@ -99,7 +105,6 @@ class FirebaseAnalyticsDriver extends IFirebaseAnalyticsDriver {
 
   AnalyticsEventItem _analyticsEventItem(Map<String, dynamic> i) {
     return AnalyticsEventItem(
-      parameters: i['parameters'],
       coupon: i['coupon']?.toString(),
       itemId: i['itemId']?.toString(),
       currency: i['currency']?.toString(),
@@ -108,10 +113,10 @@ class FirebaseAnalyticsDriver extends IFirebaseAnalyticsDriver {
       itemListId: i['itemListId']?.toString(),
       locationId: i['locationId']?.toString(),
       affiliation: i['affiliation']?.toString(),
-      index: int.tryParse(i['index'].toString()),
-      price: num.tryParse(i['price'].toString()),
       promotionId: i['promotionId']?.toString(),
       itemVariant: i['itemVariant']?.toString(),
+      index: int.tryParse(i['index'].toString()),
+      price: num.tryParse(i['price'].toString()),
       creativeName: i['creativeName']?.toString(),
       creativeSlot: i['creativeSlot']?.toString(),
       itemCategory: i['itemCategory']?.toString(),
@@ -123,6 +128,7 @@ class FirebaseAnalyticsDriver extends IFirebaseAnalyticsDriver {
       promotionName: i['promotionName']?.toString(),
       discount: num.tryParse(i['discount'].toString()),
       quantity: int.tryParse(i['quantity'].toString()),
+      parameters: _convertToMapStringObject(i['parameters']),
     );
   }
 
@@ -131,14 +137,14 @@ class FirebaseAnalyticsDriver extends IFirebaseAnalyticsDriver {
     required Map<String, dynamic> params,
   }) async {
     try {
-      final items = ((params['items'] ?? []) as List<Map<String, dynamic>>)
+      final items = List<Map<String, dynamic>>.from((params['items'] ?? []))
           .map(_analyticsEventItem)
           .toList();
       await instance.logAddToCart(
         items: items,
         currency: params['currency']?.toString() ?? 'BRL',
         value: double.tryParse(params['value'].toString()),
-        parameters: _convertToMapStringObject(params['parameters'] ?? {}),
+        parameters: _convertToMapStringObject(params['parameters']),
         callOptions: AnalyticsCallOptions(
           global: params['callOptions']?['global'] ?? false,
         ),
@@ -157,7 +163,7 @@ class FirebaseAnalyticsDriver extends IFirebaseAnalyticsDriver {
     required Map<String, dynamic> params,
   }) async {
     try {
-      final items = ((params['items'] ?? []) as List<Map<String, dynamic>>)
+      final items = List<Map<String, dynamic>>.from((params['items'] ?? []))
           .map(_analyticsEventItem)
           .toList();
       await instance.logBeginCheckout(
@@ -165,7 +171,7 @@ class FirebaseAnalyticsDriver extends IFirebaseAnalyticsDriver {
         coupon: params['coupon']?.toString(),
         currency: params['currency']?.toString() ?? 'BRL',
         value: double.tryParse(params['value'].toString()),
-        parameters: _convertToMapStringObject(params['parameters'] ?? {}),
+        parameters: _convertToMapStringObject(params['parameters']),
         callOptions: AnalyticsCallOptions(
           global: params['callOptions']?['global'] ?? false,
         ),
@@ -184,7 +190,7 @@ class FirebaseAnalyticsDriver extends IFirebaseAnalyticsDriver {
     required Map<String, dynamic> params,
   }) async {
     try {
-      final items = ((params['items'] ?? []) as List<Map<String, dynamic>>)
+      final items = List<Map<String, dynamic>>.from((params['items'] ?? []))
           .map(_analyticsEventItem)
           .toList();
       await instance.logPurchase(
@@ -196,7 +202,7 @@ class FirebaseAnalyticsDriver extends IFirebaseAnalyticsDriver {
         value: double.tryParse(params['value'].toString()),
         transactionId: params['transactionId']?.toString(),
         shipping: double.tryParse(params['shipping'].toString()),
-        parameters: _convertToMapStringObject(params['parameters'] ?? {}),
+        parameters: _convertToMapStringObject(params['parameters']),
         callOptions: AnalyticsCallOptions(
           global: params['callOptions']?['global'] ?? false,
         ),
@@ -205,6 +211,129 @@ class FirebaseAnalyticsDriver extends IFirebaseAnalyticsDriver {
       return Right(unit);
     } catch (exception, stackTrace) {
       debugPrint('FirebaseAnalyticsDriver.logPurchase: $exception');
+      crashLog.capture(exception: exception, stackTrace: stackTrace);
+      return Left(Exception(exception));
+    }
+  }
+
+  @override
+  Future<Either<Exception, Unit>> logSelectItem({
+    required Map<String, dynamic> params,
+  }) async {
+    try {
+      final items = List<Map<String, dynamic>>.from((params['items'] ?? []))
+          .map(_analyticsEventItem)
+          .toList();
+      instance.logSelectItem(
+        items: items,
+        itemListId: params['itemListId']?.toString(),
+        itemListName: params['itemListName']?.toString(),
+        parameters: _convertToMapStringObject(params['parameters']),
+        callOptions: AnalyticsCallOptions(
+          global: params['callOptions']?['global'] ?? false,
+        ),
+      );
+      debugPrint('FirebaseAnalyticsDriver.logSelectItem $params');
+      return Right(unit);
+    } catch (exception, stackTrace) {
+      debugPrint('FirebaseAnalyticsDriver.logSelectItem: $exception');
+      crashLog.capture(exception: exception, stackTrace: stackTrace);
+      return Left(Exception(exception));
+    }
+  }
+
+  @override
+  Future<Either<Exception, Unit>> logViewItem({
+    required Map<String, dynamic> params,
+  }) async {
+    try {
+      final items = List<Map<String, dynamic>>.from((params['items'] ?? []))
+          .map(_analyticsEventItem)
+          .toList();
+      instance.logViewItem(
+        items: items,
+        currency: params['currency']?.toString() ?? 'BRL',
+        value: double.tryParse(params['value'].toString()),
+        parameters: _convertToMapStringObject(params['parameters']),
+      );
+      debugPrint('FirebaseAnalyticsDriver.logViewItem $params');
+      return Right(unit);
+    } catch (exception, stackTrace) {
+      debugPrint('FirebaseAnalyticsDriver.logViewItem: $exception');
+      crashLog.capture(exception: exception, stackTrace: stackTrace);
+      return Left(Exception(exception));
+    }
+  }
+
+  @override
+  Future<Either<Exception, Unit>> logAddPaymentInfo({
+    required Map<String, dynamic> params,
+  }) async {
+    try {
+      final items = List<Map<String, dynamic>>.from((params['items'] ?? []))
+          .map(_analyticsEventItem)
+          .toList();
+      instance.logAddPaymentInfo(
+        items: items,
+        coupon: params['coupon']?.toString(),
+        paymentType: params['paymentType']?.toString(),
+        currency: params['currency']?.toString() ?? 'BRL',
+        value: double.tryParse(params['value'].toString()),
+        parameters: _convertToMapStringObject(params['parameters']),
+        callOptions: AnalyticsCallOptions(
+          global: params['callOptions']?['global'] ?? false,
+        ),
+      );
+      debugPrint('FirebaseAnalyticsDriver.logAddPaymentInfo $params');
+      return Right(unit);
+    } catch (exception, stackTrace) {
+      debugPrint('FirebaseAnalyticsDriver.logAddPaymentInfo: $exception');
+      crashLog.capture(exception: exception, stackTrace: stackTrace);
+      return Left(Exception(exception));
+    }
+  }
+
+  @override
+  Future<Either<Exception, Unit>> logAddShippingInfo({
+    required Map<String, dynamic> params,
+  }) async {
+    try {
+      final items = List<Map<String, dynamic>>.from((params['items'] ?? []))
+          .map(_analyticsEventItem)
+          .toList();
+      instance.logAddShippingInfo(
+        items: items,
+        coupon: params['coupon']?.toString(),
+        shippingTier: params['shippingTier']?.toString(),
+        currency: params['currency']?.toString() ?? 'BRL',
+        value: double.tryParse(params['value'].toString()),
+        parameters: _convertToMapStringObject(params['parameters']),
+        callOptions: AnalyticsCallOptions(
+          global: params['callOptions']?['global'] ?? false,
+        ),
+      );
+      debugPrint('FirebaseAnalyticsDriver.logAddShippingInfo $params');
+      return Right(unit);
+    } catch (exception, stackTrace) {
+      debugPrint('FirebaseAnalyticsDriver.logAddShippingInfo: $exception');
+      crashLog.capture(exception: exception, stackTrace: stackTrace);
+      return Left(Exception(exception));
+    }
+  }
+
+  @override
+  Future<Either<Exception, Unit>> logout({
+    required String name,
+  }) async {
+    try {
+      await Future.wait([
+        instance.setUserId(id: null),
+        setUserProperty(name: name, value: ''),
+      ]);
+      debugPrint('FirebaseAnalyticsDriver logout efetuado com sucesso.');
+      return Right(unit);
+    } catch (exception, stackTrace) {
+      debugPrint('FirebaseAnalyticsDriver.logout: $exception');
       crashLog.capture(exception: exception, stackTrace: stackTrace);
       return Left(Exception(exception));
     }