+++ lib/src/data/drivers/firebase/firebase_driver.dart
@@ -1,4 +1,5 @@
-import 'package:firebase_core/firebase_core.dart' show Firebase;
+import 'package:firebase_core/firebase_core.dart'
+    show Firebase, FirebaseOptions;
 import 'package:flutter/foundation.dart';
 
 import '../../../domain/interfaces/either.dart';
@@ -7,9 +8,46 @@ import '../../../infra/drivers/firebase/i_firebase_driver.dart'
 
 class FirebaseDriver extends IFirebaseDriver {
   @override
-  Future<Either<Exception, Unit>> init() async {
+  Future<Either<Exception, Unit>> init({
+    String apiKey = '',
+    String appId = '',
+    String messagingSenderId = '',
+    String projectId = '',
+    String? authDomain,
+    String? databaseURL,
+    String? storageBucket,
+    String? measurementId,
+    String? trackingId,
+    String? deepLinkURLScheme,
+    String? androidClientId,
+    String? iosClientId,
+    String? iosBundleId,
+    String? appGroupId,
+  }) async {
     try {
-      await Firebase.initializeApp();
+      FirebaseOptions? options;
+      if (apiKey.isNotEmpty &&
+          appId.isNotEmpty &&
+          messagingSenderId.isNotEmpty &&
+          projectId.isNotEmpty) {
+        options = FirebaseOptions(
+          apiKey: apiKey,
+          appId: appId,
+          messagingSenderId: messagingSenderId,
+          projectId: projectId,
+          authDomain: authDomain,
+          databaseURL: databaseURL,
+          storageBucket: storageBucket,
+          measurementId: measurementId,
+          trackingId: trackingId,
+          deepLinkURLScheme: deepLinkURLScheme,
+          androidClientId: androidClientId,
+          iosClientId: iosClientId,
+          iosBundleId: iosBundleId,
+          appGroupId: appGroupId,
+        );
+      }
+      await Firebase.initializeApp(options: options);
       debugPrint('FirebaseDriver iniciado com sucesso.');
       return Right(unit);
     } catch (e) {