+++ lib/src/domain/services/i_sentry_crashlytics_service.dart
@@ -1,9 +1,0 @@
-import '../interfaces/either.dart';
-import 'i_crash_log_service.dart';
-
-abstract class ISentryCrashlyticsService extends ICrashLogService {
-  Future<Either<Exception, Unit>> identify({
-    required Map<String, dynamic> user,
-  });
-  Future<Either<Exception, Unit>> unidentify();
-}