+++ lib/src/infra/drivers/firebase/i_firebase_analytics_driver.dart
@@ -13,6 +13,9 @@ abstract class IFirebaseAnalyticsDriver extends IAppTrackingDriver {
     required String value,
     Map<String, Object>? params,
   });
+  Future<Either<Exception, Unit>> logout({
+    required String name,
+  });
   Future<Either<Exception, Unit>> addToCart({
     required Map<String, dynamic> params,
   });
@@ -20,6 +23,18 @@ abstract class IFirebaseAnalyticsDriver extends IAppTrackingDriver {
     required Map<String, dynamic> params,
   });
   Future<Either<Exception, Unit>> purchase({
+    required Map<String, dynamic> params,
+  });
+  Future<Either<Exception, Unit>> logSelectItem({
+    required Map<String, dynamic> params,
+  });
+  Future<Either<Exception, Unit>> logViewItem({
+    required Map<String, dynamic> params,
+  });
+  Future<Either<Exception, Unit>> logAddPaymentInfo({
+    required Map<String, dynamic> params,
+  });
+  Future<Either<Exception, Unit>> logAddShippingInfo({
     required Map<String, dynamic> params,
   });
 }