+++ lib/src/infra/drivers/i_sentry_crashlytics_driver.dart
@@ -1,9 +1,0 @@
-import '../../domain/interfaces/either.dart';
-import 'i_crash_log_driver.dart';
-
-abstract class ISentryCrashlyticsDriver extends ICrashLogDriver {
-  Future<Either<Exception, Unit>> identify({
-    required Map<String, dynamic> user,
-  });
-  Future<Either<Exception, Unit>> unidentify();
-}